+++ CHANGELOG.md
@@ -1,3 +1,10 @@
-# Change Log
+## v1.1 - 2017-03-11
+### Changed
+- Almost complete rewrite of the addon
+### Added
+- Right-Click dropdown menu with display options
+### Fixed
+- Not working to it's fullest on broker bars other than Bazooka
+
 ## v1.0 - 2017-03-07
 - Initial Release